acitivity and filtering activity added
