+++ templates/open-general-licences/open-general-licence.html
@@ -163,6 +163,6 @@
 	</dl>
 
 	<h2 class="govuk-heading-m">{% lcs 'open_general_licences.Detail.ACTIVITY' %}</h2>
-
+	{% include "filters.html" %}
 	{% include 'includes/audit-trail.html' %}
 {% endblock %}
